initial organization of the project in folders
